+++ mod_cedec/classe/Classe.DefesaAgora.php
@@ -148,7 +148,9 @@
                 chdir(PATH.'/anexo/def_civil_agora');
 		        $dirAnexo = getcwd();
 		if(!empty($id['imagem1'])){
-                        unlink($dirAnexo.'/'.$id['imagem1']);
+                        if(file_exists($dirAnexo.'/'.$id['imagem1'])){
+                            unlink($dirAnexo.'/'.$id['imagem1']);
+                        }
                 }
         
                 return true;
